Create README.md
In this project, I used the Open3D library to estimate depth information from a 2D image and convert it into 3D coordinates in the form of a point cloud. The main objective was to extract the depth and orientation of objects in an image using computer vision techniques. This was part of an internship project where we aimed to guide users in a virtual reality (VR) environment by recognizing the position and orientation of real-world objects captured in 2D images.